World & Character Basics
Replaced the previous 370World Unity project files with the ones I had been working on. (I admit I didn't look to see if you had added anything to that one already so sorry if I overwrote anything, we can always re-add it from the version history!) It has a different name, but we can update that whenever.

This includes some basic player movement, camera follow, and world generation scripts.

Currently each tile of the world has a 50% chance of being null or being a gray square, just as proof of concept for basic randomization.