Added Alternate Cursor Mode & Improvements to World Controller + Tiles
Implemented an alternate cursor mode (crosshair as opposed to tile selection) which can be used for aiming weapons. Game starts with crosshair mode set, press 2 to change to tile select and 1 to swap back.

Implemented a Dictionary to keep track of Tile and GameObject pairs. This involved removing that funny lambda bit so we don't have to worry about that anymore.

Cleaned up singleton pattern code of WorldController, moving to Awake() and putting in a collapsible #region for cleanliness! ;)