Started work on Animal AI and added utility method to Tile class
Got started on basic AI for our animals – for now our test animal just walks around then waits for a bit and picks a new spot to wander off to.

Also added a useful function to our Tile class which takes another Tile as an argument and returns true if they are neighbors. An addition bool parameter can specify whether or not to check for diagonal neighbors as well.