Updated AnimalAI so that animals will no longer walk on water.
The animal will check each tile ahead of it (within a short distance) and if it finds one that is not walkable it will set a new target position to wander to.
This involved modifying Tile.cs to have an isWalkable boolean and WorldController's OnTileTypeChanged set set it true or false.

Added another utility to Tile.cs for checking if a tile has walkable neighbors – I was thinking about using this for something but then I realized it was a terrible idea lol. We can remove it later if it is never used but maybe it will come in handy.

Also moved the Singleton region to the top of WorldController so it looks nicer.

Thinking about renaming the AnimalAI.cs script to just Animal.cs but I don't know.