Added our first test turret!
Added a test turret which detects anything that enters its attack range and, if it's an enemy, shoots at them after a short delay.

Made a prefab of our test animal and our test turret as well as one for the turret's bullets.

Added "Enemy" and "Bullet" tags so that objects can be differentiated on collision or entering a trigger.

Made sure our player and animal each had a collider and rigidbody as well.