Created base Destructible class for anything that has health, takes damage, and is destroyed
Includes variables to be set for max health, hit & death sfx, and items to drop when destroyed.

Can be applied to animals, enemies, and resource nodes alike. I've already applied it to our tree, rock, and test animal prefabs. The test animal currently dies after 3 shots of the turret.

Known issue: The turret prints some null reference exceptions to the console if its target dies while it is still trying to lock on or shoot at it.