Upgraded our PlaySound() function
Previously we had two PlaySound() functions where the second was an overloaded version to play the sound in world space (meaning it would play more from left or right speaker depending on player/camera position).

I've improved upon this by combining the two so that we now only have one PlaySound() function that with an optional parameter for world space position.

Also added a third parameter (following the AudioClip to play and the optional position Vector3) which is an optional bool indicating whether or not you want the sound clip to play pitch shifted or not. Simply include "true" as the last paremeter when you call PlaySound() to have the pitch randomly adjusted. For now I am using this with the player's footsteps to add some variation when moving around.