Started work on Inventory and Items
Created a UI inventory bar on the bottom of the screen and started work on inventory functionality. Created a class for inventory items and implemented a placeholder item for testing.

The player can walk over the items on the ground to pick them up and add them to the inventory, unless it is already full. An appropriate sound effect is played depending on whether the item was picked up or not.