Installable walls and turrets
TEST_Turret and TEST_Wall now start out as uninstalled inventory items so that the player can walk over and pick them up. Once in the inventory (in slot 1/2/3) the player can press 1/2/3 key to equip the item and if it is installable, their cursor will change to the tile select cursor and when they click the object will be installed. Walls are stackable in the inventory while turrets are not and installation should be working properly for both.

To-do:
- Find a way to differentiate between wall items to pick up & walls that are installed. Maybe different sprites or a lil icon?
- Fix input issue where sometimes you have to press the 1/2/3 key more than once to get it to select the item
- Copy/paste code for 1/2/3 keys so that all 7 slots have full functionality
- Make sure only 1 object can be installed to each Tile a time