Integrated new turret art assets!
Had to update the Turret script to rotate the head to face the target. I think it looks great already for now but feel free to tweak anything you think can be improved!

Still can start shooting too early (as in as soon as lock on timer ends but still not yet rotated to aim at enemy) but not really an issue just a stylistic choice if we want to change that.